requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
gradio==4.8.0

# Computer Vision & Video Processing
opencv-python==4.8.1.78
mediapipe==0.10.8
ultralytics==8.0.196
numpy==1.24.3
pillow==10.1.0

# Audio Processing
pydub==0.25.1
openai-whisper==20231117
praat-parselmouth==0.4.3
librosa==0.10.1

# NLP & AI
google-generativeai==0.3.2
transformers==4.35.2
torch==2.1.1

# Data Processing
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Database & Storage
pymongo==4.6.0
motor==3.3.2

# Report Generation
reportlab==4.0.7
weasyprint==60.2

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
pydantic-settings==2.1.0

# NLP & Text Processing
nltk==3.8.1 